+++ packages/api-group/src/deleteGroupInvitation.ts
@@ -1,0 +1,4 @@
+import { makeApiUrl, deleteRequest } from '@gtms/api-common'
+
+export const deleteGroupInvitationAPI = (id: string) =>
+  deleteRequest(makeApiUrl(`groups/invitations/${id}`))