+++ packages/state-group/src/groupInvitations/groupInvitations.store.ts
@@ -1,0 +1,18 @@
+import { IGroupInvitation } from '@gtms/commons/models'
+import { EntityState, EntityStore } from '@datorama/akita'
+
+export interface GroupInvitationsState
+  extends EntityState<IGroupInvitation, number> {
+  offset: number
+  total: number
+}
+
+export class GroupInvitationsStore extends EntityStore<GroupInvitationsState> {
+  constructor() {
+    super(undefined, {
+      name: 'groupInvitations',
+    })
+  }
+}
+
+export const groupInvitationsStore = new GroupInvitationsStore()