+++ packages/app-andrew/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.31.1](https://github.com/gtms-org/gtms-frontend/compare/@gtms/app-andrew@0.31.0...@gtms/app-andrew@0.31.1) (2020-06-18)
+
+
+### Bug Fixes
+
+* GEOT-285: add group invitation tabs in group settings ([#657](https://github.com/gtms-org/gtms-frontend/issues/657)) ([9cc4628](https://github.com/gtms-org/gtms-frontend/commit/9cc462889ea36cac1aa11d74b86ede1ba20c83f9))
+
+
+
+
+
 # [0.31.0](https://github.com/gtms-org/gtms-frontend/compare/@gtms/app-andrew@0.30.0...@gtms/app-andrew@0.31.0) (2020-06-17)